create pokedex with no css
