+++ week4/src/pages/pokesearch.tsx
@@ -1,30 +1,6 @@
-import React, {ChangeEvent, FormEvent, SetStateAction, useState} from 'react'
-
-interface searchProps{
-    search: string;
-    change: (event: ChangeEvent<HTMLInputElement>) => void;
-    submit: (event: FormEvent<HTMLFormElement>) => void;
-}
-
-interface displayProps{
-    display: {name: string,
-        id: string,
-        type: string[],
-        picture: string
-    };
-}
-
-const Search : React.FC<searchProps> = ({search, change, submit}) => 
-{
-    
-    return (
-        <form onSubmit = {submit}>
-            <input type = "text" placeholder = "Search Pokemon" value = {search} onChange = {change} />
-            <button type = "submit">Search</button>
-        </form>
-    )
-
-}
+import React, {ChangeEvent, FormEvent, useState} from 'react'
+import Search from './components/search'
+import DisplayTable from './components/displaytable'
 
 function PokeSearch()
 {
@@ -48,9 +24,6 @@ function PokeSearch()
         fetch(`https://pokeapi.co/api/v2/pokemon/${search}`).then(data => data.json()
         )
         .then(function(pokemon : Record<string, any>){
-            // pokemon.types.forEach((object : Record<string, any>) => {
-            //     console.log(object.type)
-            // })
             setDisplay(prevState => ({
                 ...prevState,
                 name: pokemon.name,
@@ -69,35 +42,5 @@ function PokeSearch()
         </>
     )
 }
-
-const DisplayTable : React.FC<displayProps> = ({display}) =>
-{
-    const {name, id, type, picture} = display
-    return (
-        <>
-        <div>
-            {name}
-        </div>
-        <div>
-            {id}
-        </div>
-        <div>
-        {type.map((pokemon, index) => (
-        <div key={index}>
-          <p>{pokemon}</p>
-        </div>
-        ))}
-        </div>
-        <div>
-            <img src={picture}></img>
-        </div>
-        </>
-    )
-}
-
-// function DisplayPokemon()
-// {
-
-// }
 
 export default PokeSearch
